eliminando registros de base de datos por medio de Resource Controller (delete) y tambien con Restful controller

+++ app/views/noticias/index.blade.php
@@ -1,5 +1,7 @@
-<?php $estado=\Illuminate\Support\Facades\Session::get('estado'); ?>
+{{HTML::script('http://code.jquery.com/jquery-2.1.4.min.js');}}
+{{HTML::script('js/restfulizer.js');}}
 
+<?php $estado=\Illuminate\Support\Facades\Session::get('estado'); ?>
 @if ($estado == 'salvar')
     La noticia se guardó con éxito <br>
 @endif
@@ -11,9 +13,11 @@
 
     <li>
         <a href="{{ route('noticias.show', $noticia->id) }}" title="">
-            {{$noticia ->titulo}} </a>  -
-    {{ Form::open(array('method'=>'DELETE','route'=>array('noticias.destroy', $noticia->id))) }}
-    {{ Form::submit('Eliminar', array('class'=>'Clase boton')) }}
-    {{ Form::close() }}
+            {{$noticia ->titulo}} </a> - <a href="{{ route('noticias.destroy', $noticia->id) }}"
+data-method="delete" rel="nofollow" data-confirm="¿Está seguro que desea eliminar la noticia?">Borrar</a>
+
+        {{ Form::open(array('method'=>'DELETE','route'=>array('noticias.destroy', $noticia->id))) }}
+        {{ Form::submit('Eliminar', array('class'=>'Clase boton')) }}
+        {{ Form::close() }}
     </li>
 @endforeach
